Polishing, S01E02
- Added back `display_decode_error` because used by `downloader.cpp`
- Rework `Downloader` to use `int64_t`

+++ src/download/downloader.hpp
@@ -21,7 +21,7 @@ namespace fur::download {
 
 /// A downloaded piece for a torrent file.
 struct Downloaded {
-  size_t index;
+  int64_t index;
   std::vector<uint8_t> content;
 };
